整理代码，检测代码

+++ src/com/dave/controller/ResultController.java
@@ -11,6 +11,7 @@ import javax.servlet.http.HttpServletResponse;
 import org.apache.poi.ss.usermodel.Workbook;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
@@ -21,7 +22,7 @@ import com.dave.entity.ResultQues;
 import com.dave.service.ResultService;
 
 /**
- * 调查结果Controller
+ * Result控制层
  * 
  * @author Dave 20190909
  *
@@ -31,8 +32,10 @@ import com.dave.service.ResultService;
 public class ResultController {
 	@Autowired
     private ResultService resultService;
+	
 	/**
 	 * 调查结果管理页面
+	 * 
 	 * @return system/result_list
 	 */
     @RequestMapping("doResultListUI")
@@ -40,58 +43,77 @@ public class ResultController {
 		return "system/result_list";
 	}
     /**
-	 * 详细用户调查结果
+	 * 详细调查结果
+	 * 
 	 * @return system/result_detail
 	 */
     @RequestMapping("doResultDetailUI")
 	public String doResultDetailUI() {
 		return "system/result_detail";
 	}
+    
 	/**
-	 * 查找调查结果数据
+	 * 查找查询所有调查结果
+	 * 
 	 * @param pageCurrent
-	 * @param resultName
+	 * @param paperName
+	 * @param startDate
+	 * @param endDate
 	 * @return
 	 */
     @RequestMapping("doFindResultList")
     @ResponseBody
-    public JsonResult doFindResultList(int pageCurrent, String paperName, String startDate, String endDate) {
+    public JsonResult doFindResultList(Integer pageCurrent, String paperName, String startDate, String endDate) {
+    	if(pageCurrent == null || pageCurrent <= 0) {
+    		return new JsonResult("当前页面数不能为空");
+    	}
     	PageObject<Result> list = resultService.findResultList(pageCurrent, paperName, startDate, endDate);
     	return new JsonResult(list);
     }
+    
     /**
      * 删除调查结果
+     * 
      * @param resultIds
      * @return
      */
     @RequestMapping("doDeleteResult")
     @ResponseBody
     public JsonResult doDeleteResult(Integer... resultIds) {
+    	if(StringUtils.isEmpty(resultIds) || resultIds.length <= 0) {
+    		return new JsonResult("调查结果ID集不能为空");
+    	}
     	int row = resultService.deleteResult(resultIds);
     	if(row > 0) {
     		return new JsonResult("Delete Succeed!", row); 		
     	}
     	return new JsonResult("Delete failed!!");
     }
+    
     /**
-     * 删除调查结果
+     * 跟进调查结果ID获取调查结果问题
+     * 
      * @param resultIds
      * @return
      */
     @RequestMapping("doGetResultQues")
     @ResponseBody
-    public JsonResult doGetResultQues(int resultId) {
+    public JsonResult doGetResultQues(Integer resultId) {
+    	if(StringUtils.isEmpty(resultId)) {
+    		return new JsonResult("问题ID不能为空");
+    	}
     	List<ResultQues> list = resultService.getResultQues(resultId);
     	return new JsonResult(list);
     }
+    
     /**
-	 * 导出调查结果
-	 * 
-	 * @param response
-	 * @param paperName
-	 * @param startDate
-	 * @param endDate
-	 */
+     * 以Excel报表导出调查结果
+     * 
+     * @param response
+     * @param paperName
+     * @param startDate
+     * @param endDate
+     */
 	@RequestMapping("doExportResult")
 	public void doExportResult(HttpServletResponse response, String paperName, String startDate, String endDate) {
 		try {
@@ -109,4 +131,5 @@ public class ResultController {
 			e.printStackTrace();
 		}
 	}
+	
 }
